pyproject.toml
[project]
name = "spintrak"
version = "0.1.0"
readme = "README.md"
requires-python = "==3.12.3"
dependencies = [
"absl-py==2.3.1",
"aiofiles==24.1.0",
"annotated-types==0.7.0",
"antlr4-python3-runtime==4.9.3",
"anyio==4.10.0",
"argon2-cffi==25.1.0",
"argon2-cffi-bindings==25.1.0",
"arrow==1.3.0",
"asttokens==3.0.0",
"async-lru==2.0.5",
"attrs==25.3.0",
"audioread==3.0.1",
"av==15.0.0",
"babel==2.17.0",
"beautifulsoup4==4.13.5",
"bleach==6.2.0",
"blis==1.3.0",
"brotli==1.1.0",
"catalogue==2.0.10",
"certifi==2025.8.3",
"cffi==1.17.1",
"charset-normalizer==3.4.3",
"click==8.2.1",
"cloudpathlib==0.21.1",
"cloudpickle==3.1.1",
"colorlog==6.9.0",
"comm==0.2.3",
"confection==0.1.5",
"contourpy==1.3.3",
"cycler==0.12.1",
"cymem==2.0.11",
"debugpy==1.8.17",
"decorator==5.2.1",
"defusedxml==0.7.1",
"demucs==4.0.1",
"docopt==0.6.2",
"dora-search==0.1.12",
"einops==0.8.1",
"encodec==0.1.1",
"executing==2.2.1",
"fast-jl==0.1.3",
"fastapi==0.116.1",
"fastjsonschema==2.21.2",
"ffmpy==0.6.1",
"filelock==3.13.1",
"flashy==0.0.2",
"fonttools==4.60.0",
"fqdn==1.5.1",
"fsspec==2024.6.1",
"gradio==5.43.1",
"gradio-client==1.12.1",
"groovy==0.1.2",
"grpcio==1.74.0",
"h11==0.16.0",
"hf-xet==1.1.8",
"httpcore==1.0.9",
"httpx==0.28.1",
"huggingface-hub==0.34.4",
"hydra-colorlog==1.2.0",
"hydra-core==1.3.2",
"idna==3.10",
"ipykernel==6.30.1",
"ipython==9.5.0",
"ipython-pygments-lexers==1.1.1",
"ipywidgets==8.1.7",
"isoduration==20.11.0",
"jedi==0.19.2",
"jinja2==3.1.4",
"joblib==1.5.1",
"json5==0.12.1",
"jsonpointer==3.0.0",
"jsonschema==4.25.1",
"jsonschema-specifications==2025.9.1",
"julius==0.2.7",
"jupyter==1.1.1",
"jupyter-client==8.6.3",
"jupyter-console==6.6.3",
"jupyter-core==5.8.1",
"jupyter-events==0.12.0",
"jupyter-lsp==2.3.0",
"jupyter-server==2.17.0",
"jupyter-server-terminals==0.5.3",
"jupyterlab==4.4.7",
"jupyterlab-pygments==0.3.0",
"jupyterlab-server==2.27.3",
"jupyterlab-widgets==3.0.15",
"kiwisolver==1.4.9",
"lameenc==1.8.1",
"langcodes==3.5.0",
"language-data==1.3.0",
"lark==1.2.2",
"lazy-loader==0.4",
"librosa==0.11.0",
"lightning-utilities==0.15.2",
"llvmlite==0.44.0",
"marisa-trie==1.3.0",
"markdown==3.8.2",
"markdown-it-py==4.0.0",
"markupsafe==2.1.5",
"matplotlib==3.10.6",
"matplotlib-inline==0.1.7",
"mdurl==0.1.2",
"mistune==3.1.4",
"mpmath==1.3.0",
"msgpack==1.1.1",
"murmurhash==1.0.13",
"nbclient==0.10.2",
"nbconvert==7.16.6",
"nbformat==5.10.4",
"nest-asyncio==1.6.0",
"networkx==3.3",
"notebook==7.4.5",
"notebook-shim==0.2.4",
"num2words==0.5.14",
"numba==0.61.2",
"numpy==2.1.2",
"nvidia-cublas-cu12==12.8.4.1",
"nvidia-cuda-cupti-cu12==12.8.90",
"nvidia-cuda-nvrtc-cu12==12.8.93",
"nvidia-cuda-runtime-cu12==12.8.90",
"nvidia-cudnn-cu12==9.10.2.21",
"nvidia-cufft-cu12==11.3.3.83",
"nvidia-cufile-cu12==1.13.1.3",
"nvidia-curand-cu12==10.3.9.90",
"nvidia-cusolver-cu12==11.7.3.90",
"nvidia-cusparse-cu12==12.5.8.93",
"nvidia-cusparselt-cu12==0.7.1",
"nvidia-ml-py==13.580.82",
"nvidia-nccl-cu12==2.27.3",
"nvidia-nvjitlink-cu12==12.8.93",
"nvidia-nvtx-cu12==12.8.90",
"omegaconf==2.3.0",
"openunmix==1.3.0",
"orjson==3.11.2",
"packaging==25.0",
"pandas==2.3.2",
"pandocfilters==1.5.1",
"parso==0.8.5",
"pexpect==4.9.0",
"pillow==11.0.0",
"platformdirs==4.3.8",
"pooch==1.8.2",
"preshed==3.0.10",
"prometheus-client==0.23.1",
"prompt-toolkit==3.0.52",
"protobuf==6.32.0",
"psutil==7.1.0",
"ptyprocess==0.7.0",
"pure-eval==0.2.3",
"pycparser==2.22",
"pydantic==2.11.7",
"pydantic-core==2.33.2",
"pydub==0.25.1",
"pygments==2.19.2",
"pyparsing==3.2.5",
"python-dateutil==2.9.0.post0",
"python-json-logger==3.3.0",
"python-multipart==0.0.20",
"pytz==2025.2",
"pyyaml==6.0.2",
"pyzmq==27.1.0",
"referencing==0.36.2",
"regex==2025.7.34",
"requests==2.32.5",
"retrying==1.4.2",
"rfc3339-validator==0.1.4",
"rfc3986-validator==0.1.1",
"rfc3987-syntax==1.1.0",
"rich==14.1.0",
"rpds-py==0.27.1",
"ruff==0.12.9",
"safehttpx==0.1.6",
"safetensors==0.6.2",
"scikit-learn==1.7.1",
"scipy==1.16.1",
"seaborn==0.13.2",
"semantic-version==2.10.0",
"send2trash==1.8.3",
"sentencepiece==0.2.1",
"setuptools==80.9.0",
"shellingham==1.5.4",
"six==1.17.0",
"smart-open==7.3.0.post1",
"sniffio==1.3.1",
"soundfile==0.13.1",
"soupsieve==2.8",
"soxr==0.5.0.post1",
"spacy==3.8.7",
"spacy-legacy==3.0.12",
"spacy-loggers==1.0.5",
"srsly==2.5.1",
"stack-data==0.6.3",
"starlette==0.47.2",
"submitit==1.5.3",
"sympy==1.13.3",
"tabulate==0.9.0",
"tensorboard==2.20.0",
"tensorboard-data-server==0.7.2",
"terminado==0.18.1",
"thinc==8.3.6",
"threadpoolctl==3.6.0",
"tinycss2==1.4.0",
"tokenizers==0.21.4",
"tomlkit==0.13.3",
"torch==2.8.0",
"torchaudio==2.8.0",
"torchmetrics==1.8.1",
"torchtext==0.18.0",
"torchvision==0.23.0",
"tornado==6.5.2",
"tqdm==4.67.1",
"traitlets==5.14.3",
"traker==0.3.2",
"transformers==4.55.3",
"treetable==0.2.5",
"triton==3.4.0",
"typer==0.16.1",
"types-python-dateutil==2.9.0.20250822",
"typing-extensions==4.12.2",
"typing-inspection==0.4.1",
"tzdata==2025.2",
"uri-template==1.3.0",
"urllib3==2.5.0",
"uvicorn==0.35.0",
"wasabi==1.1.3",
"wcwidth==0.2.13",
"weasel==0.4.1",
"webcolors==24.11.1",
"webencodings==0.5.1",
"websocket-client==1.8.0",
"websockets==15.0.1",
"werkzeug==3.1.3",
"wheel==0.45.1",
"widgetsnbextension==4.0.14",
"wrapt==1.17.3",
"xformers==0.0.32.post2"
]

[tool.uv.extra-build-dependencies]
fast-jl = ["torch"]

[project.scripts]
spintrak = "spintrak.main:app"
experiment_runner = "spintrak.experiments.experiment_runner:app"
experiment_1_results = "spintrak.experiments.experiment1_generate_results_graph:run"
experiment_3_results = "spintrak.experiments.experiment3_generate_results_graph:run"
experiment_4_results = "spintrak.experiments.experiment4_generate_results_graph:run"

[build-system]
requires = ["uv_build>=0.8.19,<0.9.0", "torch"]
build-backend = "uv_build"


